Made changes to rh_sticky_answers so that they are addressed correctly.
Long term I would like to remove all
references to main:: except for those
in the init routines.  The references
can be replaced by a locally defined
pointer to the  value in the current safe
compartment.

This procedure should make it easier to
maintain the macros.  All references to
global variables will occur in the init
routine.

--Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@1462 c0722133-6baf-4dd8-8699-98d999cd4f06