This is a preliminary attempt at adding answer checkers that use the parser.  If you have a parser object (like a formula or a vector), you can use the ->cmp method to get an answer checker appropriate for the object.  For example:
	 Context('Vector');
	 $v = Vector(1,2,3);
	 #  print text of problem
	 ANS($v->cmp);

would produce an answer checker that matched against the vector <1,2,3>.
Most of the parameters are set in the Context, but additional ones can
be passed to the checker (this still needs work).

See the sample problems for more examples.

The checkers for Reals, Complexes, Points, Vectors, Matrices and
Intervals work pretty well.  The ones for List and Union objects need
more work, as they only do an ORDERED check at the moment (the student
answers must be in the same order).  This needs to be fixed.
The checker for Formula objects (which corresponds to fun_cmp) still
needs to be written.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2593 c0722133-6baf-4dd8-8699-98d999cd4f06