Fixes for handling intervals better.  In particular, deal better with ambiguous situations, like (1,2) that could be a point.
You can use [a,a] to specify the single point a.  I'd like to make {a}
represent this (and maybe even {a,b,c} to form finite sets), but this
would interfere with the use of { and } as alternative parentheses.
(Does anyone really use these?)

Do we need a cross product for intervals (and sets) as a means of
representing regions in R^2 (or higher)?


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2594 c0722133-6baf-4dd8-8699-98d999cd4f06