The List object's answer checker is now considerably improved.  It will allow you to specify an unorderd or ordered check, and whether to print various kinds of hints for wrong answers.
The checker can return partial credit, but WW doesn't deal well with
this at the moment (some small changes to Problem.pm can take care of
that).  Partial credit should be based on $showPartialCorrectAnswers,
but it is disabled for now.

The list answer checker probably still needs some work, especially
with non-numeric lists.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2599 c0722133-6baf-4dd8-8699-98d999cd4f06