Changed the string used for the student answer to be taken from the unevaluated formula rather than the fully-evaluated one.  This leaves constants unreplaced, for one thing.
Currently the parser combines constants automatically during the parse
[which should be controllable by a flag], so student answers that
produce numbers are fully evaluated at this point already, except for
constant substitution.  This could be changed so that even for
numberic results, the student sees the parsed expression (rather than
the numeric result), which might help them identify where they and the
parser differ in their interpretation.  It is still nice to see the
final computed result, however, but there is no place to show the
parsed input as well as the final result.  One solution would be to
add another column (wasting space).  Another would be to make the
value shown be "parsed-value = computed-value" when they are not the
same thing.  A bit ugly, but works within the current framework.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2634 c0722133-6baf-4dd8-8699-98d999cd4f06