Added easy calls to disable/enable collections of functions or individual functions in the current context.  So you can say
	   Parser::Context::Functions::Disable('Trig');

to disallow the use of trig functions, or

	   Parser::Context::Functions::Disable('Trig','Numeric');
	   Parser::Context::Functions::Enable('sqrt');

to disable trig and numeric functions (like log, exp, abs, etc) but
allow sqrt.

Note that disabling sqrt() does not disable x^(1/2) (this must be done
through disabling the ^ and ** operators), and disabling abs() does
not disable |x| (which must be done by removing the '|' definition
from the Context's parens list).


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2672 c0722133-6baf-4dd8-8699-98d999cd4f06