Added ability to have linear adaptive parameters in the function answer checker.  It only works for real-valued functions, though. To use a parameter, use
       Context()->variables->add(a=>'Parameter');

and then use 'a' as a variable within your answer.  The student will
not be allowed to enter the parameter, but the professor will.

Note that the correct answer will show the full professor's answer,
including the parameters, even though the student can't type it.  Is
this the right behaviour?


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2687 c0722133-6baf-4dd8-8699-98d999cd4f06