Added context flags to control automatic reduction of operations that involve only constant values.  The default is to combine constants, but this can now be disabled.
Also added a new flag to control the addition of extra parentheses
that make string values painfully unambiguous (these are for
situations that students tend to get wrong).  By default this is
turned off (so that professor answers won't contain excessive
parentheses), but the answer checker will be made to turn this on for
student answers.

Finally, when extra parens are added to formulas, they alternate
between ( ) and [ ] to make it easier to read.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2789 c0722133-6baf-4dd8-8699-98d999cd4f06