Updated i(), j() and k() functions to check if the Parser is loaded, and if not, return a value that will cause perl to report a better error than calling the parser outside of its normal context.
Since the current answer checkers call perl directly to evaluate their
expressions, when a variable is used by a student that is not a
variable used by the checker, this can cause perl to try to parse
those variables as function calls, or other perl constructs (like
matches if you use 'm' or translations if you use 'y').  This is bad
(and is not a problem with the new parser).

In particular, if the student uses 'k' in a formula for which k is not
defined, for example, perl will try to call the k() function.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@2932 c0722133-6baf-4dd8-8699-98d999cd4f06