Update the overloaded operators so that they can be overridden by subclasses of the predefined object classes.  This involves calling the objects method rather than using a hard reference to the routine in the parent class.
Also, change the name of the string comparison routine to
compare_string to avoid conflicts with cmp that is used to produce the
answer checker for the class.

Finally, in Value.pm, promotePrecedence no longer has to do fancy
footwork to get "special" precedence to work (this was a hack to get
around the misfeature of the overloaded operators -- now that that is
being handled correctly, there is no need for it).


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3192 c0722133-6baf-4dd8-8699-98d999cd4f06