Changed loadMacros() to use search for amcros files via a user-cusomizable path of directories.  This is specified in the global.conf file, and can include any number of directories.
The path can include "." to mean look in the .pg file's directory.
This makes it possible to store special macros files with the code
that uses them.

By default, the path is ".", the course/templates/macro directory,
then the pg/macros directory (in that order).  This duplicates the
current behaviour, except for the ".", which is new.

You must also make the updates to the global.conf,
webwork2/lib/WeBWorK/PG.pm, and webwork2/lib/WeBWorK/PG/Local/pm files.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3225 c0722133-6baf-4dd8-8699-98d999cd4f06