Fixed conceptual error in compare_vec_solution.  This should fix bug #670.  In my opinion the entire concept of vec_solution_cmp should be reconsidered.  In solving an underdetermined linear equation of the form Ax-b=0 it seems to me that the solutions answer in the form: x= a +bt+cu+ds  where a,b,c,d are vectors should simply be evaluated to see if it satisfies Ax-b=0 for 5 or six values of a,b,c,d  -- checking the solution should use a vector valued version of fun_cmp.
As it is, the student's coefficients for a,b,c,d are compared with the
instructors to see if they span the same space. This is quite a bit more
complicated -- and indeed the method came up with the wrong answer.

I believe I have the method corrected, but I would suggest that this
answer evaluator be replaced with one which operates more
directly and is therefore easier to maintain.  Am I missing something in
this analysis?  Has someone else created answer evaluators for this type
of problem?

 -- Mike


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3314 c0722133-6baf-4dd8-8699-98d999cd4f06