Modified the perl() method to make calles to Value:: directly rather than to the main:: stubs for creating Value objects.  This will make it work better from within packages or when Parser.pl hasn't been loaded.
A few things still won't work in this case; e.g., calls to Closed()
for intervals (these could be handled better using ->with()) and to
functions like Factorial and log10 that are defined in Parser.pl.

Finally, there should be better object-based control over what
routines are called to create these objects, so that subclasses of the
Value objects will be able to be generated correctly.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3352 c0722133-6baf-4dd8-8699-98d999cd4f06