Fix use of $SIG{__WARN__} and $SIG{__DIE__}.
* Each change to $SIG{__WARN__} and $SIG{__DIE__} is now dynamically
scoped with "local", rather than changing globally and then restoring at
the end of the block.

* The special value "DEFAULT" is used instead of sub {CORE::die(@_)} or
sub {CORE::warn(@_)} when preparing to eval STRING code.

* Where $SIG{__WARN__} and $SIG{__DIE__} are overridden to do
preprocessing with PG_errorMessage, the previously installed handlers
will be called instead of the built-in warn or die functions after
preprocessing occurs. (For example, the warn and die handlers defined in
Apache::WeBWorK or the warn handler defined in WeBWorK::PG will be
called.)

* The behavior in process_answers is modified so that the custom
handlers are installed once before the answer evaluation loop. This
causes them to be in effect during the setup/teardown code, but I don't
think this will be a problem.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3426 c0722133-6baf-4dd8-8699-98d999cd4f06