Moved the union and set reduction checks from cmp_equal to cmp_compare so that the messages can be produced even in lists of unions.
Improved the message facilities for the list comparison so that
cmp_compare can make errors that refer to the number of the entry in
the student's answer (otherwise an extra "There was a problem with
your nth value" message needs to be prepended, which looks bad if it
can be avoided).

Use the "extra" answer checker to report syntax error messages in unordered
lists.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3504 c0722133-6baf-4dd8-8699-98d999cd4f06