Added a parserCustomization.pl file that is loaded whenever the Parser.pl file is.  This can be used on a course-by-course basis to customize features of the Parser.  Some examples are listed in the file.
You need to be VERY careful when doing this, as this will affect ALL
problems that use the Parser.  Your changes may cause some .pg files
to fail if they rely on the default behavior that you are changing.

If you wish to override settings in one of the predefined contexts
(like "Vector"), then define $context{Vector} as a copy of
$Parser::Context::Default::context{Vector}, as in the commented out
examples.  Then Context("Vector") will use your copy rather than the
original.  If you were to change the version in
Parser::Context::Default, that would change the parser for ALL
courses, not just yours.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3588 c0722133-6baf-4dd8-8699-98d999cd4f06