Movified the copying of functions from Complex1:: into main:: to avoid conflicts with the PGcommonFunctions.pl versions (these errors were trapped, but still show up in the error log unnecessarily).
Also commented out some code that was not doing anything other than
producing error messages in the error log.  (It was left over from a
syntax check on the professor's answer, but the actual check was
removed, leaving a portion that tries to process the answer, but
usually fails (because things like "1+4i" need to be converted to
"1+4*i" before they can be used in PG_answer_eval, but that was not
being done).

Because of this, it is not possible currently to do cplx_cmp("1+4i"),
and instead you must to cplx_cmp(new Complex(1,4)).

To fix this, you would need to call check_syntax (and the other
filters that are called on the student's answer) before calling
PG_answer_eval.  Of course, you should only do this when the
professor's answer isn't already a Complex object.

I am going to work on a Legacy module like the ones for num_cmp and
fun_cmp to replace cplx_cmp, which should avoid these problems and
make the changes suggested above unnecessary.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3601 c0722133-6baf-4dd8-8699-98d999cd4f06