Roll back Mike's change, since this can cause unwanted error messages in list comparisons and other situations where the equality check could compare different types of objects (for example) or where the comparison could legitimately fail.  We only want to report errors here when the user-supplied checker code fails.
This change would cause errors to be reported, for example, when the
user (incorrectly) entered a vector one time and then a formula
(correctly) the next time.  When the previous answer is checked against
the current one, this change would report an error rather than
silently ignoring the error, as it should.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3754 c0722133-6baf-4dd8-8699-98d999cd4f06