Take care of the issue where Preview could be used to gather information about whether an answer is correct by using the "this answer is the same as the previous one" message.  Now if Preview is used, the warning is given whether the answer is right or not.
(Maybe it is not necessary to check for correct at all, since that was
there to prevent the message from being issued when there are multiple
answer blanks and the student is working on other ones.  But the exact
match check will take care of those.  Should the message be issued if
the student changes a correct answer to an equivalent one?)


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3757 c0722133-6baf-4dd8-8699-98d999cd4f06