Really fix fun_cmp equivalence warning message.
We show the equivalence message when: we're not in preview mode AND the
answers are equivalent AND the answers are not identical. We DON'T CARE
whether the answers are correct or not, because that leaks information
in multipart questions when $showPartialCorrectAnswers is off.

Both the Parser-version (FUNCTION_CMP) and the non-Parser-version
(ORIGINAL_FUNCTION_CMP) behave as descibed above now.

And this time we test *before* we commit...


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@3814 c0722133-6baf-4dd8-8699-98d999cd4f06