Made str_cmp accept blank answers again.  Fixes bug #1123.
Note:  if you use blank answers, the "n of the questions remain
unanswered" message will tell you how many of the answers should be
left blank.  It is probably worth working out a method of having blank
entries that are only marked correct when all the other answers are
correct (this would take a special grader; I have this as part of the
weighted grader, but it is not common knowledge).


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@4626 c0722133-6baf-4dd8-8699-98d999cd4f06