This begins some major updates to the Parser and Value libraries. This may make the MathObjects unstable as things shake down, but I'll try to keep that to a minimum.
This update reorganizes how the overloading of operations are
implemented.  The base Value class is now overloaded rather than the
individual Value classes.  This makes it easier to make subclasses
that take advantage of the overloading.

The code that handles promoting one MathObject to another has been
updated as a part of this.  This should also make subclassing the
existing MathObjects more flexible.

In the past, Formula objects included a pointer to the Context in
which the object was created, but other MathObject did not.  This
update is the first round of updates to include the context in the
other MathObjects, but more work needs to be done.

The changes here should not affect existing problems; it is possible
that some macro files will need to be updated, but none of the
standard ones in pg/macros seem to need it at the moment.


git-svn-id: http://svn.webwork.maa.org/system/trunk/pg@4975 c0722133-6baf-4dd8-8699-98d999cd4f06